move drawer to separate component
